getting rid of warnings

+++ trunk/base/flow_objects.c
@@ -181,8 +181,8 @@ struct GridExportList : GridObject {
 };
 
 GRID_INLET(GridExportList,0) {
-	int n = in->dim->prod();
-	if (n>1000000) RAISE("list too big (%d elements, max 1000000)", n);
+	long n = in->dim->prod();
+	if (n>1000000) RAISE("list too big (%ld elements, max 1000000)", n);
 	list = rb_ary_new2(n+2);
 	this->n = n;
 	rb_ivar_set(rself,SI(@list),list); // keep
